Use wrappers.proto
Use protobuf wrappers to be explicit about optional values. By default, all fields in protobuf 3 are optional. If a field is missing in the message, the fields are set to a default value, e.g. `boolean flag = false`. This means the the receiving side does not know if the field is missing or explictly set to false. Therefore, we will use protobuf wrappers that will use optional types of the specific language. As an example, if the field is missing, the value in Scala will be `val flag: Option[Boolean] = None` and if the boolean is explictly set to `false` then the value is in Scala is `val flag: Option[Boolean] = Some(false)`.